+++ app/src/main/java/woowacourse/shopping/data/cart/model/AddCartRequestBody.kt
@@ -1,3 +1,3 @@
-package woowacourse.shopping.data.cart
+package woowacourse.shopping.data.cart.model
 
 data class AddCartRequestBody(val productId: Int)
