+++ app/src/main/java/woowacourse/shopping/data/cart/model/UpdateQuantityRequestBody.kt
@@ -1,3 +1,3 @@
-package woowacourse.shopping.data.cart
+package woowacourse.shopping.data.cart.model
 
 data class UpdateQuantityRequestBody(val quantity: Int)
