+++ app/src/main/java/woowacourse/shopping/data/common/model/BaseResponse.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.common
+package woowacourse.shopping.data.common.model
 
 data class BaseResponse<T>(
     val message: String,
